requirements.txt
Django
djangorestframework
mysqlclient
phonenumber_field
phonenumbers
dj_rest_auth